microcopy_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
openai==1.95.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sentencepiece==0.2.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
